+++ lib/service/Gateway.dart
@@ -2,7 +2,7 @@ import 'package:http/http.dart';
 import 'package:illinois/model/StudentCourse.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Config.dart';
-import 'package:illinois/service/DeepLink.dart';
+import 'package:rokwire_plugin/service/deep_link.dart';
 import 'package:rokwire_plugin/service/network.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/service.dart';
@@ -11,8 +11,6 @@ import 'package:rokwire_plugin/utils/utils.dart';
 class Gateway with Service implements NotificationsListener {
 
   static const String notifyBuildingDetail = "edu.illinois.rokwire.gateway.building.detail";
-
-  List<Uri>? _deepLinkUrisCache;
 
   // Singleton Factory
 
@@ -25,9 +23,8 @@ class Gateway with Service implements NotificationsListener {
   @override
   void createService() {
     NotificationService().subscribe(this, [
-      DeepLink.notifyUri,
+      DeepLink.notifyUiUri,
     ]);
-    _deepLinkUrisCache = <Uri>[];
     super.createService();
   }
 
@@ -37,11 +34,6 @@ class Gateway with Service implements NotificationsListener {
     super.destroyService();
   }
 
-
-  @override
-  void initServiceUI() {
-    _processCachedDeepLinkUris();
-  }
 
   @override
   Set<Service> get serviceDependsOn {
@@ -92,33 +84,9 @@ class Gateway with Service implements NotificationsListener {
   static String get buildingDetailUrl => '${DeepLink().appUrl}/gateway/building_detail';
 
   void _onDeepLinkUri(Uri? uri) {
-    if (uri != null) {
-      if (_deepLinkUrisCache != null) {
-        _cacheDeepLinkUri(uri);
-      } else {
-        _processDeepLinkUri(uri);
-      }
-    }
-  }
-
-  void _processDeepLinkUri(Uri uri) {
-    if (uri.matchDeepLinkUri(Uri.tryParse(buildingDetailUrl))) {
-      NotificationService().notify(notifyBuildingDetail, uri.queryParameters.cast<String, dynamic>());
-    }
-  }
-
-  void _cacheDeepLinkUri(Uri uri) {
-    _deepLinkUrisCache?.add(uri);
-  }
-
-  void _processCachedDeepLinkUris() {
-    if (_deepLinkUrisCache != null) {
-      List<Uri> deepLinkUrisCache = _deepLinkUrisCache!;
-      _deepLinkUrisCache = null;
-
-      for (Uri deepLinkUri in deepLinkUrisCache) {
-        _processDeepLinkUri(deepLinkUri);
-      }
+    if ((uri != null) && uri.matchDeepLinkUri(Uri.tryParse(buildingDetailUrl))) {
+      try { NotificationService().notify(notifyBuildingDetail, uri.queryParameters.cast<String, dynamic>()); }
+      catch (e) { print(e.toString()); }
     }
   }
 
@@ -126,8 +94,8 @@ class Gateway with Service implements NotificationsListener {
 
   @override
   void onNotification(String name, dynamic param) {
-    if (name == DeepLink.notifyUri) {
-      _onDeepLinkUri(param);
+    if (name == DeepLink.notifyUiUri) {
+      _onDeepLinkUri(JsonUtils.cast(param));
     }
   }